Moved InlineError into separate component
